Major refactor of code in the data-provider component.
