+++ components/data-exporter/src/error.rs
@@ -1,19 +1,19 @@
 #[derive(Debug)]
 pub enum Error {
-    ResponseError(icu_data_provider::ResponseError),
-    PayloadError(icu_data_provider::PayloadError),
+    ResponseError(icu_data_provider::error::ResponseError),
+    PayloadError(icu_data_provider::error::PayloadError),
     SerializerError(erased_serde::Error),
     IoError(std::io::Error),
 }
 
-impl From<icu_data_provider::ResponseError> for Error {
-    fn from(err: icu_data_provider::ResponseError) -> Error {
+impl From<icu_data_provider::error::ResponseError> for Error {
+    fn from(err: icu_data_provider::error::ResponseError) -> Error {
         Error::ResponseError(err)
     }
 }
 
-impl From<icu_data_provider::PayloadError> for Error {
-    fn from(err: icu_data_provider::PayloadError) -> Error {
+impl From<icu_data_provider::error::PayloadError> for Error {
+    fn from(err: icu_data_provider::error::PayloadError) -> Error {
         Error::PayloadError(err)
     }
 }
