+++ components/data-provider-json/Cargo.toml
@@ -14,30 +14,16 @@ include = [
     "README.md"
 ]
 
-[features]
-default = ["std"]
-std = ["no-std-compat/std", "serde/std", "serde_json/std"]
-
 [dependencies]
 icu-data-provider = { path = "../data-provider" }
 icu-locale = { path = "../locale" }
 
-[dependencies.async-trait]
-version = "0.1.30"
-
 [dependencies.serde]
 version = "1.0"
-default-features = false
-features = ["derive", "alloc"]
+features = ["derive"]
 
 [dependencies.serde_json]
 version = "1.0"
-default-features = false
-features = ["alloc"]
-
-[dependencies.no-std-compat]
-version = "0.4.0"
-features = ["alloc"]
 
 [[test]]
 name = "file_io"
