+++ components/data-provider-json/src/schema.rs
@@ -1,12 +1,9 @@
-use icu_data_provider as datap;
+use icu_data_provider::structs;
 use serde::{Deserialize, Serialize};
-
-#[allow(unused_imports)]
-use std::prelude::v1::*;
 
 #[derive(Serialize, Deserialize, Debug)]
 pub(crate) struct DecimalJsonSchema {
-    pub(crate) symbols_v1_a: datap::decimal::SymbolsV1,
+    pub(crate) symbols_v1_a: structs::decimal::SymbolsV1,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
