+++ components/data-provider/src/cloneable_any.rs
@@ -1,5 +1,3 @@
-use std::prelude::v1::*;
-
 use downcast_rs::impl_downcast;
 use downcast_rs::Downcast;
 use std::fmt::Debug;
