+++ components/data-provider/src/structs/mod.rs
@@ -1,0 +1,2 @@
+pub mod decimal;
+pub mod plurals;
