+++ components/data-provider/src/structs/plurals.rs
@@ -1,13 +1,13 @@
 // Plural types
 
-use std::prelude::v1::*;
-
+use crate::data_key::Category;
+use crate::data_key::DataKey;
 use serde::{Deserialize, Serialize};
 use std::any::TypeId;
 use std::borrow::Cow;
 
-pub(crate) fn get_type_id(data_key: &crate::DataKey) -> Option<TypeId> {
-    if data_key.category != crate::Category::Plurals {
+pub(crate) fn get_type_id(data_key: &DataKey) -> Option<TypeId> {
+    if data_key.category != Category::Plurals {
         return None;
     }
     match data_key.sub_category.as_str() {
